+++ crates/syn/src/types/trident_fuzz_test_executor.rs
@@ -1,8 +1,0 @@
-use syn::{Generics, Ident, Type};
-
-#[derive(Debug)]
-pub struct TridentFuzzTestExecutor {
-    pub ident: Ident,
-    pub generics: Generics,
-    pub client_type: Type,
-}