+++ crates/template/src/test_fuzz.rs
@@ -34,13 +34,16 @@ impl Template {
 
             #(#use_statements)*
 
-            #[derive(Default, FuzzTestExecutor)]
-            struct FuzzTest {
-                client: TridentSVM,
+            #[derive(Default)]
+            struct FuzzTest<C> {
+                client: C,
             }
 
             #[flow_executor]
-            impl FuzzTest {
+            impl<C: FuzzClient + std::panic::RefUnwindSafe> FuzzTest<C> {
+                fn new(client: C) -> Self {
+                    Self { client }
+                }
                 #[init]
                 fn start(&mut self) {
                     #(#programs)*